Rename rules package
